backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.4.2
python-multipart==0.0.6
google-generativeai==0.3.2
python-dotenv==1.0.0